Opus review, code cleanup
